src/attribute_analyzer.py
"""Logic for inferring caller attributes using Azure Cognitive Services."""
from typing import Optional

import azure.cognitiveservices.speech as speechsdk

from .caller_attributes import CallerAttributes


class CallerAttributeAnalyzer:
    """Analyze audio to infer caller attributes."""

    def __init__(self, speech_key: str, speech_region: str):
        self._speech_key = speech_key
        self._speech_region = speech_region

    def analyze(self, audio_path: str) -> CallerAttributes:
        """Return inferred attributes for the audio file at *audio_path*."""
        speech_config = speechsdk.SpeechConfig(
            subscription=self._speech_key,
            region=self._speech_region,
        )
        audio_config = speechsdk.audio.AudioConfig(filename=audio_path)
        recognizer = speechsdk.SpeechRecognizer(
            speech_config=speech_config,
            audio_config=audio_config,
        )

        # Placeholder for real attribute analysis using Azure APIs.
        # Azure SDK does not directly provide gender or age estimation; this
        # typically uses Speech Analytics or custom models. This example shows
        # where such logic would be integrated.
        result = recognizer.recognize_once()

        attributes = CallerAttributes()
        if result.reason == speechsdk.ResultReason.RecognizedSpeech:
            # TODO: integrate with Azure Speech Analytics or additional services
            # to infer gender, age_group and emotional_state from result or
            # additional calls.
            pass

        return attributes

