requirements.txt
azure-cognitiveservices-speech
fastapi
jinja2
uvicorn
python-dotenv
